+++ optimal_scheduler/rootfs/etc/services.d/example/Abstraction/Asset_types/Consumers/EVCharger/EVCharger.py
@@ -1,9 +1,7 @@
 # EVCharger class
-import os
 
 from AbsConsumer import AbsConsumer
 from utils.utils import createClass
-
 
 class EVCharger(AbsConsumer):
 
@@ -12,8 +10,8 @@ class EVCharger(AbsConsumer):
         super().__init__(configuration, name)
         self.simul = createClass("SIMU_" + name, configuration['Simulate'])
 
-    def doSimula(self, **kwargs):
-        self.config['calendar'] = kwargs['calendar']
+    def doSimula(self, calendar, **kwargs):
+        self.config['calendar'] = calendar
         return self.simul.simula(self.simul, self.config, kwargs_simulation=kwargs['kwargs_simulation'])
 
     def canviaSimula(self, simImpl):  # simImpl will be an implementation of the Simulate interface