+++ optimal_scheduler/rootfs/etc/services.d/example/Abstraction/Asset_types/Generators/PV/PV.py
@@ -6,9 +6,13 @@ from utils.utils import createClass
 class PV(AbsGenerator):
 
     def __init__(self, configuration, name):
+
         super().__init__(configuration, name)
         self.simul = createClass("SIMU_" + name, configuration['Simulate'])
 
     def doSimula(self, calendar, **kwargs):
-        self.config['calendar'] = kwargs['calendar']
+        self.config['calendar'] = calendar
         return self.simul.simula(self.simul, self.config, kwargs_simulation=kwargs['kwargs_simulation'])
+
+    def canviaSimula(self, simImpl):  # simImpl will be an implementation of the Simulate interface
+        self.simul = simImpl